Add sentry suffix to issue URL
This adds the "sentry" suffix to the URL built to link to an issue. As
we have upgraded to a new sentry, the URL_PREFIX option changed, along
with the location of several options in sentry (such as organizations).
This commit makes it so that the link can point to the right location,
which is the sentry location under the main sentry URL.